# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-Commerce Inventory & Sales System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: lowStock
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: customerId
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithDetails"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (auto-reduces stock)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with line items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"
        "404":
          description: Not found
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithDetails"

  # ─── Inventory Logs ──────────────────────────────────────────────────────────
...
  /inventory/update:
    post:
      operationId: updateInventory
      tags: [inventory]
      summary: Manually update product stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # ─── Analytics ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        category:
          type: string
        price:
          type: number
        stockQuantity:
          type: integer
        description:
          type: string
        isLowStock:
          type: boolean
      required: [productId, productName, category, price, stockQuantity, description, isLowStock]
    CreateProductBody:
      type: object
      properties:
        productName:
          type: string
        category:
          type: string
        price:
          type: number
        stockQuantity:
          type: integer
        description:
          type: string
      required: [productName, category, price, stockQuantity, description]
...
    OrderItem:
      type: object
      properties:
        orderDetailId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        subtotal:
          type: number
        price:
          type: number
      required: [orderDetailId, productId, productName, quantity, subtotal, price]
    OrderWithDetails:
      type: object
      properties:
        orderId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        orderDate:
          type: string
        totalAmount:
          type: number
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [orderId, customerId, customerName, orderDate, totalAmount, status, items]
    CreateOrderBody:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineInput"
      required: [customerId, items]
    OrderLineInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
      required: [status]
...
    UpdateInventoryBody:
      type: object
      properties:
        productId:
          type: integer
        newStock:
          type: integer
      required: [productId, newStock]
...