# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Inventory & Sales System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLineInput } from "./orderLineInput";

export interface CreateOrderBody {
  customerId: number;
  items: OrderLineInput[];
}