# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Inventory & Sales System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  productName: string;
  category: string;
  price: number;
  stockQuantity: number;
  description: string;
}