# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Inventory & Sales System API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  orderDetailId: number;
  productId: number;
  productName: string;
  quantity: number;
  subtotal: number;
  price: number;
}