# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Inventory & Sales System API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  productId: number;
  productName: string;
  category: string;
  price: number;
  stockQuantity: number;
  description: string;
  isLowStock: boolean;
}